# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ComicWise API
  description: |
    API for the ComicWise comic reading platform.

    ## Authentication
    Most endpoints require authentication via NextAuth session cookies.
    Admin endpoints require the user to have `role: "admin"`.

    ## Rate Limiting
    API routes are rate-limited to prevent abuse.

  version: 1.0.0
  contact:
    name: ComicWise Team
  license:
    name: MIT

servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://comicwise.app
    description: Production server

tags:
  - name: Health
    description: Health check endpoints
  - name: Comics
    description: Comic browsing and search
  - name: Chapters
    description: Chapter content
  - name: Auth
    description: Authentication endpoints
  - name: Bookmarks
    description: User bookmarks
  - name: Admin
    description: Admin management (requires admin role)

paths:
  /api/health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns application health status
      operationId: getHealth
      responses:
        "200":
          description: Application is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
        "503":
          description: Application is unhealthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /api/auth/session:
    get:
      tags:
        - Auth
      summary: Get current session
      description: Returns the current user session if authenticated
      operationId: getSession
      responses:
        "200":
          description: Session data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"

  /api/comics:
    get:
      tags:
        - Comics
      summary: List comics
      description: Get a paginated list of comics with optional filters
      operationId: getComics
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [Ongoing, Hiatus, Completed, Dropped, "Season End", "Coming Soon"]
        - name: genre
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of comics
          content:
            application/json:
              schema:
                type: object
                properties:
                  comics:
                    type: array
                    items:
                      $ref: "#/components/schemas/Comic"
                  pagination:
                    $ref: "#/components/schemas/Pagination"

  /api/comics/{slug}:
    get:
      tags:
        - Comics
      summary: Get comic by slug
      description: Get detailed information about a specific comic
      operationId: getComicBySlug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comic details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComicDetail"
        "404":
          description: Comic not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/upload:
    post:
      tags:
        - Admin
      summary: Upload file
      description: Upload an image file (requires authentication)
      operationId: uploadFile
      security:
        - sessionAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                folder:
                  type: string
                  description: Target folder for the upload
      responses:
        "200":
          description: Upload successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/dev/seed:
    post:
      tags:
        - Admin
      summary: Seed database
      description: Seed the database with test data (development only)
      operationId: seedDatabase
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: Seeding completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "403":
          description: Forbidden in production
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-seed-api-key
      description: API key for seed endpoint

  schemas:
    HealthStatus:
      type: object
      required:
        - status
        - timestamp
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
        timestamp:
          type: string
          format: date-time
        version:
          type: string
        uptime:
          type: integer
          description: Server uptime in seconds
        checks:
          type: object
          properties:
            database:
              type: string
              enum: [ok, error]
            cache:
              type: string
              enum: [ok, error]

    Session:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        expires:
          type: string
          format: date-time

    User:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        image:
          type: string
          format: uri
        role:
          type: string
          enum: [user, admin, moderator]

    Comic:
      type: object
      required:
        - id
        - title
        - slug
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        coverImage:
          type: string
          format: uri
        status:
          type: string
          enum: [Ongoing, Hiatus, Completed, Dropped, "Season End", "Coming Soon"]
        rating:
          type: number
          format: float
        chapterCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    ComicDetail:
      allOf:
        - $ref: "#/components/schemas/Comic"
        - type: object
          properties:
            authors:
              type: array
              items:
                $ref: "#/components/schemas/Author"
            artists:
              type: array
              items:
                $ref: "#/components/schemas/Artist"
            genres:
              type: array
              items:
                $ref: "#/components/schemas/Genre"
            chapters:
              type: array
              items:
                $ref: "#/components/schemas/ChapterSummary"

    Author:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        bio:
          type: string
        image:
          type: string
          format: uri

    Artist:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        bio:
          type: string
        image:
          type: string
          format: uri

    Genre:
      type: object
      required:
        - id
        - name
        - slug
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string

    ChapterSummary:
      type: object
      required:
        - id
        - number
        - title
      properties:
        id:
          type: integer
        number:
          type: number
        title:
          type: string
        createdAt:
          type: string
          format: date-time

    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string

    ActionResult:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: object
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string